+++ components/ServiceCard.tsx
@@ -1,7 +1,7 @@
 "use client"
 import React from 'react';
 import {Card, CardHeader, CardBody, Image} from "@nextui-org/react";
-import { ArrowIcon, WebIcon, SeoIcon, DevIcon } from './svg/work-exp';
+import { ArrowIcon, WebIcon, SeoIcon, DevIcon } from './svg/serviceIcons';
 
 import jsIcon from "../public/service-icons/pointer.svg";
